renamed SoundType
